added server with /landmarks endpoint
