+++ UrbServer/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "urbserver",
+  "version": "1.0.0",
+  "description": "Server for Urbscope",
+  "main": "app.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node app.js"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.16.3",
+    "request": "^2.85.0"
+  }
+}
