+++ UrbServer/results.json
@@ -1,0 +1,74 @@
+{
+    "landmarks": [
+        {
+            "destinationID": "4cac512344a8224bbe1a3240",
+            "name": "The Debate",
+            "latitude": 51.50113791186581,
+            "longitude": -0.12478416977691376,
+            "address": [
+                "Portcullis House",
+                "City of Westminster",
+                "Greater London",
+                "United Kingdom"
+            ],
+            "categoryID": "4bf58dd8d48988d16d941735",
+            "category": "Café"
+        },
+        {
+            "destinationID": "4d91d34a033ba35d5dbd770d",
+            "name": "Parliament Jubilee Café",
+            "latitude": 51.500176067103936,
+            "longitude": -0.12606761624060436,
+            "address": [
+                "United Kingdom"
+            ],
+            "categoryID": "4bf58dd8d48988d16d941735",
+            "category": "Café"
+        },
+        {
+            "destinationID": "4c167c976470a5936de756d2",
+            "name": "Roux at Parliament Square",
+            "latitude": 51.501225413834284,
+            "longitude": -0.12818236675651487,
+            "address": [
+                "11 Great George St",
+                "London",
+                "Greater London",
+                "SW1P 3AD",
+                "United Kingdom"
+            ],
+            "categoryID": "4bf58dd8d48988d10c941735",
+            "category": "French Restaurant"
+        },
+        {
+            "destinationID": "4f60dcbfe5e8d4902df2c210",
+            "name": "Gillray's Steakhouse & Bar",
+            "latitude": 51.501824010633726,
+            "longitude": -0.11982061234039655,
+            "address": [
+                "County Hall, Westminster Bridge Road",
+                "London",
+                "Greater London",
+                "SE1 7PB",
+                "United Kingdom"
+            ],
+            "categoryID": "4bf58dd8d48988d1cc941735",
+            "category": "Steakhouse"
+        },
+        {
+            "destinationID": "4c532062f5f3d13a196cfbf7",
+            "name": "Zen China Restaurant",
+            "latitude": 51.502058765191684,
+            "longitude": -0.11976850603052665,
+            "address": [
+                "County Hall, Riverside Building (Westminster Bridge Road)",
+                "London",
+                "Greater London",
+                "SE1 7PB",
+                "United Kingdom"
+            ],
+            "categoryID": "4bf58dd8d48988d145941735",
+            "category": "Chinese Restaurant"
+        }
+    ]
+}
